Modifications according to the code review
